% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives.R
\docType{class}
\name{Alternatives' attributes-class}
\alias{Alternatives' attributes-class}
\alias{ob_alternatives_att}
\title{ob_alternatives_att}
\arguments{
\item{N}{The number of decision makers}

\item{p}{The number of attributes of each decision makers}
}
\description{
The class of observed decision makers' attributes
}

\examples{
J <- 10; q <- 3; a <- ob_alternatives_att$new(J=J, q=q);
          a$gen("student", param=list(location=rep(100, 3), df=3)); a$gen("help")

}
