% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_design.R
\name{alternatives_gen}
\alias{alternatives_gen}
\title{Generate experimental design (choice sets)}
\usage{
alternatives_gen(experimental_design, n = NULL, seed = NULL, format = "long")
}
\arguments{
\item{experimental_design}{Input experimental design configuration}

\item{n}{It is possible to generate different number of choice sets without directly modifying the `experimental_design` object.
User may locally overwrite the `n` value for the particular execution.
This parameter is for internal use in conjunction with various wrappers.}

\item{seed}{The seed to be set before attempting to generate data.
Defaults to NULL.}

\item{format}{The resulting data format specification.
At this tage only the "long" format is supported.
Meaning each line contains one alternative.}
}
\value{
data.frame A data.frame (Z) with simulated choice sets
(one row per alternative if élongé format was declared).
}
\description{
Generate experimental design data (2 in standart notation) from an `experimental_design` object.
}
\examples{
# Create alternatives
alt1 <- alternative$new()
alt1$add_attributes(Quality = purrr::rbernoulli(p = 0.6), Price = rnorm(mean = 5))
alt2 <- alternative$new()
alt2$add_attributes(Size = runif(min = 0, max = 1), Price = rnorm(mean = 6))

# Regroup alternatives into design
edesign <- experimental_design$new(alternatives = list(alt1, alt2))
Z <- alternatives_gen(edesign)
}
