% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{best_choice}
\alias{best_choice}
\title{best_choice}
\description{
The method of the Experiment class which computes the optimal decision makers' choice
}
\examples{
DM_att_names <- list("X1", "X2", "X3")
AT_att_names <- list("Z1", "Z2", "Z3")
AT_names <- list("good1", "good2")
groups <- c(10, 20)
FD <- Experiment(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,groups=groups)
FD$gen_AT_attributes()
FD$gen_DM_attributes()
FD$gen_preference_coefficients()
FD$utility()
FD$Epsilon; FD$U; FD$choice_order
FD$utility("student", mu=3, df=4) # Some examples
FD$utility("discrete_uniform") # Some examples
FD$Epsilon
FD$U
FD$choice_order
FD$Epsilon
FD$U
FD$choice_order
FD$choice
}
