% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{call_design}
\alias{call_design}
\title{call_design}
\usage{
call_design(
  name,
  X,
  Z,
  U,
  choice_set_size,
  J,
  no_choice,
  nb_questions,
  format = "long"
)
}
\arguments{
\item{name}{The name of the chosen experimental design}

\item{X}{matrix of decision makers data}

\item{Z}{matrix of alternatives data}

\item{U}{Utility matrix}

\item{choice_set_size}{the number of alternative per choice set}

\item{J}{number of alternatives (no_choice included)}

\item{no_choice}{TRUE FALSE indicator}
}
\description{
The function which associates the designs' name to a function which generates it.
}
