% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{categorization}
\alias{categorization}
\title{categorization}
\usage{
categorization(X, nb_levels)
}
\arguments{
\item{X}{The matrix data}

\item{np_levels}{The vector of the number of clusters to build in each column.}
}
\value{
$clustered The cluster's average of each attribute of each alternative
$category The cluster's number of each attribute of each alternative
}
\description{
The function which classify by k-means the values of each column
 of X. Then, the values of each column is set to the center of its cluster.
}
\examples{
X <- matrix(rnorm(30), ncol=3)
nb_levels <- c(3, 2, 2)
X_clustered <- categorization(X, nb_levels)$clustered
X_category <- categorization(X, nb_levels)$category
X_clustered; X_category
}
