% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_helpers.R
\name{check_attributes}
\alias{check_attributes}
\title{Attributes check}
\usage{
check_attributes(Z, attributes)
}
\arguments{
\item{Z}{Input data.frame Z}

\item{attributes}{Desired vector of attributes to be present in Z.}
}
\value{
Z With missing attributes filled with NA's.
}
\description{
Verification that all desired characteristics are present in $Z$.
This is an internal function which serves to simplify concatenation of multiple $Z$ when alternatives have different attributes.
}
\examples{
Z <- check_attributes(Z, attributes)
}
