% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_identical}
\alias{compose_identical}
\title{Compose identical choice sets}
\usage{
compose_identical(experimental_design, n = NULL, size)
}
\arguments{
\item{experimental_design}{An `experimental_design` object.}

\item{n}{Number of different choice situations per individual.
Is NULL by default, ensuring that the function will retreive the n value from the `experimental_design`.}

\item{size}{Population size for Z to be adjusted to.}
}
\value{
data.frame Z dataset
}
\description{
This is an internal function which creates identical choice sets configurations for give experimental design.
The function simplifies the size adjustment of Z depending on the population size of X.
}
\examples{
Z <- compose_identical(experimental_design)
}
