% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Dcriterion}
\alias{Dcriterion}
\title{Dcriterion}
\usage{
Dcriterion(experimental_design, DM_att_names, AT_att_names, choice_set_size)
}
\arguments{
\item{experimental_design}{The design matrix}
}
\value{
The D-score of the experimental design

Data Frame of a full factorial design
}
\description{
The function which computes the determinant of the information matrix.
}
