% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision makers.R
\docType{class}
\name{Decision makers' attributes-class}
\alias{Decision makers' attributes-class}
\alias{ob_decision_makers_att}
\title{ob_decision_makers_att}
\arguments{
\item{N}{The number of decision makers}

\item{p}{The number of attributes of each decision makers}
}
\description{
The class of observed decision makers' attributes
}

\examples{
N <- 10; p <- 3; a <- ob_decision_makers_att$new(N=N, p=p);
          a$gen("student", param=list(location=rep(100, 3), df=3)); a$gen(); a$gen("help")

}
