% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_rule.R
\docType{class}
\name{decision_rule}
\alias{decision_rule}
\title{Decision rule class}
\description{
The `decision_rule` R6 class serves to define individuals'decision rules.
The object stores information about the decision rule:
a list of formulas to be applied on the data;
the associated random noise terms,
transformation rules and choice selection pattern.

Append a list of `call` noise definitions to the noise field.
The noise is declared as randomisation function (ex: `evd::rgumbel(loc = 0, scale = 1)`)

Append a list of `call` formula definitions to the formula field.
The formulas should be defined in the same order as alternatives the individual will face (ex: `Age + 2*Quality, 1.5*Age + Quality^2`)

The transformation should be a function of Total Utility (TU).
The default transforamtion is `expr(exp(TU) / sum(exp(TU)))`

The choice rule represents the criteria of individual's final choice.
The desired criteria should be declared as function.
The default value is `max()` (assuming the individual chooses the alternative with higher associated probability).
}
\section{Modifying methods}{
NA
}

\examples{
drule <- decision_rule$new()
drule$add_noise(evd::rgumbel(loc = 0, scale = 1), evd::rgumbel(loc = 0, scale = 2))
drule$add_formulas(Age + 2 * Quality, 1.5 * Age + Quality^2)
drule$modify_transformation(TU)
drule$modify_choice(min())
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{characteristics}}{A list of characteristics' definitions}

\item{\code{decision_rule}}{A decision rule object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{characteristics}}{A list of characteristics' definitions}

\item{\code{decision_rule}}{A decision rule object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_noise}{\code{decision_rule$add_noise()}}
\item \href{#method-add_formulas}{\code{decision_rule$add_formulas()}}
\item \href{#method-modify_transformation}{\code{decision_rule$modify_transformation()}}
\item \href{#method-modify_choice}{\code{decision_rule$modify_choice()}}
\item \href{#method-clone}{\code{decision_rule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_noise"></a>}}
\if{latex}{\out{\hypertarget{method-add_noise}{}}}
\subsection{Method \code{add_noise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decision_rule$add_noise(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_formulas"></a>}}
\if{latex}{\out{\hypertarget{method-add_formulas}{}}}
\subsection{Method \code{add_formulas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decision_rule$add_formulas(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modify_transformation"></a>}}
\if{latex}{\out{\hypertarget{method-modify_transformation}{}}}
\subsection{Method \code{modify_transformation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decision_rule$modify_transformation(transformation)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modify_choice"></a>}}
\if{latex}{\out{\hypertarget{method-modify_choice}{}}}
\subsection{Method \code{modify_choice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decision_rule$modify_choice(choice)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decision_rule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
