% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{design}
\alias{design}
\title{design}
\arguments{
\item{name}{The name of the experimental design to implement.}

\item{choice_set_size}{The number of alternative per choice set.}

\item{clustered}{Determines the way the data is represented in the experimental design table.
0 if the matrices of decision makers' characteristics and alternatives' attributes are row.
1 if the matrices of decision makers' characteristics and alternatives' attributes are clustered.
2 if the matrices of decision makers' characteristics and alternatives' attributes are categorized.}

\item{format}{If "long" (default) the design is expressed in long format and wide format otherwise.}
}
\value{
an Experimental Design as well as a some pieces of information such as the D-score, defined as the determinant of the covariance matri<- of the preference parameter (a good D-score should be small), and an estimation of the preference parameters.
}
\description{
The method of the Experiment class which generate a full factorial design
}
\examples{
DM_att_names <- list("X1", "X2", "X3")
AT_att_names <- list("Z1", "Z2", "Z3")
AT_names <- list("good1", "good2", "good3", "good4")
groups <- c(10, 20)
FD <- Experiment(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names, groups=groups)
FD$gen_AT_attributes()
FD$gen_DM_attributes()
FD$gen_preference_coefficients()
FD$utility()
FFD <- FD$design(choice_set_size=2, clustered=0) # generation of the full factorial design with row data
#by default, name="FuFD", choice_set_size = nb_alternatives
FFD1 <- FD$design(name="FuFD",choice_set_size=2, clustered=1, nb_levels_DM=c(3, 3, 4, 2), nb_levels_AT=c(3, 2, 2, 4), format="wide") # generation of the full factorial design with glustered data
FFD2 <- FD$design(choice_set_size=2, clustered=2, nb_levels_DM=c(2, 3, 4, 2), nb_levels_AT=c(2, 2, 2, 2)) # generation of the full factorial design with categorical data
FFD3 <- FD$design(name="FrFD", choice_set_size=2, clustered=2, nb_levels_DM=c(2, 3, 4, 2), nb_levels_AT=c(2, 2, 2, 2), nb_questions = 2) # Generation a a random fractional factorial design with categorical data
FFD4 <- FD$design(name="FrFD", choice_set_size=2, clustered=2, nb_levels_DM=c(2, 3, 4, 2), nb_levels_AT=c(3, 3, 3, 3), nb_questions = 2, format="wide") # Yet, we want to express this design in wide format
}
