% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_f.R
\name{designs_f}
\alias{designs_f}
\title{Design matrix generation for factors}
\usage{
designs_f(experimental_design, n = NULL, resample = NULL)
}
\arguments{
\item{experimental_design}{Input experimnetal design object}

\item{n}{The n is not used in current version}

\item{resample}{Logical.
Declares whether the dataset should be resampled (shuffled) in case of factorial design element presence.}
}
\value{
data.frame A dataset of choice sets (Z) respecting FF Design.
}
\description{
Create experimental design based on factors.
The design configuration assumes that all alternatives' attributes are "factors":
vectors declared using `c()` assignement.
Other attribute types will be ignored.
}
\examples{
# Create alternatives
alt1 <- alternative$new()
alt1$add_attributes(Quality = c(0:1), Price = c(2, 2.5))
alt2 <- alternative$new()
alt2$add_attributes(Size = c(1:3), Price = c(1.5, 2))

# Regroup alternatives into design
edesign <- experimental_design$new(alternatives = list(alt1, alt2), design = "factorial")
Z <- designs_f(edesign, n = 2)
}
