% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\title{Experiment execution}
\usage{
experiment(population, experimental_design, seed = NULL)
}
\arguments{
\item{population}{A `population` object.}

\item{experimental_design}{An `experimental_desing` object.}

\item{seed}{The desired seed to be set before data generation.
No seed is set by default (`seed = NULL`).}
}
\value{
data.frame A complete experimental dataset XZ
}
\description{
This function simulates the experimental procedure.
The function calls on `experiment_run` procedure at the first step.
Then the decision procedure is carried out.
The transformation functions are applied to TU (functions output) and then the choice criteria is applied over the results.
}
\examples{
XZ <- experiment_run(population, experimental_design, seed = 10)
}
