% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_compose.R
\name{experiment_compose}
\alias{experiment_compose}
\title{Compose experiment}
\usage{
experiment_compose(population, experimental_design, seed = NULL)
}
\arguments{
\item{population}{A `population` object.}

\item{experimental_design}{An `experimental_desing` object.}

\item{seed}{The desired seed to be set before data generation.
No seed is set by default (`seed = NULL`).}
}
\value{
data.frame A complete experimental dataset XZ
}
\description{
This function simulates the data preparation step for an experiment.
The function performs coupling between X (population) and Z (experimental_design) datasets.
}
\examples{
XZ <- experiment_run(population, experimental_design, seed = 10)
}
