% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_run.R
\name{experiment_run}
\alias{experiment_run}
\title{Run experiment}
\usage{
experiment_run(population, experimental_design, seed = NULL, XZ = NULL)
}
\arguments{
\item{population}{A `population` object.}

\item{experimental_design}{An `experimental_desing` object.}

\item{seed}{The seed to be preset for data simulation purposes.
The default is NULL to use system defined or random seed.}

\item{XZ}{The experimental data.frame to be used instead of generated one}
}
\value{
data.frame A complete experimental dataset with associated utilities.
}
\description{
This function simulates running of an experiment.
First, the `experiment_compose` is executed to obtain a coupling between X and Z matrices.
Then Deterministic Utilities (DU)  and Total Utilities (TU) are calculated using the specified formulas.
Note, that even the TU abbreviation is used this still makes possible to recalibrate the generation functions, tranforms and decision tules to create a Regret Minimisation (RM) framework i.e.
}
\examples{
# Generate individual profile and population
ind3 <- individual$new()
ind3$add_characteristics(Age = rnorm(mean = 50, sd = 4), Salary = runif(min = 1, max = 5))
drule <- decision_rule$new()
drule$add_noise(rnorm(), rnorm(sd = 2))
drule$add_formulas(Age + 2 * Quality, 1.5 * Age + Quality^2)
ind3$add_decision_rule(drule)
pop <- population$new(profiles = list(ind3), n = list(5))

# Create alternatives and regroup them into design
alt1 <- alternative$new()
alt1$add_attributes(Quality = runif(min = 0, max = 1), Price = rnorm(mean = 5))
alt2 <- alternative$new()
alt2$add_attributes(Size = runif(min = 0, max = 1), Price = rnorm(mean = 6))
edesign <- experimental_design$new(alternatives = list(alt1, alt2), n = 4)

# Run experiment
res <- experiment_run(pop, edesign)
}
