% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_design.R
\docType{class}
\name{experimental_design}
\alias{experimental_design}
\title{Eperimental design class}
\description{
The `experimental_design` R6 class regroups different alternatives.
It contains all the essential information of the desired experimental design.
It serves as a wrapper for simultaneous interactions with alternative profiiles.
The object is used for choice sets generation procedures.

Add new alternative to the list of available alternatives.

Set a new `design` configuration.

Get a vector of available attributes' names across all alternatives.

Get specified design.
}
\section{Modifying methods}{
NA
}

\section{Querrying methods}{
NA
}

\examples{
# Create alternatives
alt1 <- alternative$new()
alt1$add_attributes(Quality = purrr::rbernoulli(p = 0.6), Price = rnorm(mean = 5))
alt2 <- alternative$new()
alt2$add_attributes(Size = runif(min = 0, max = 1), Price = rnorm(mean = 6))

# Regroup alternatives into design
edesign <- experimental_design$new(alternatives = list(alt1, alt2))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alternatives}}{A list of alternatives to use for experimental design construction.}

\item{\code{design}}{An experimental design specification.
This field assumes that a savvy user may desire to extend the number of available experimental designs.
In the default configuration available designs are: "random", "factorial" and "mixed".}

\item{\code{n}}{A number of choice sets per individual for "random" design configuration.}

\item{\code{identical}}{Logical.
Declares whether the choice sets should be identical across individuals or not.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{experimental_design$new()}}
\item \href{#method-add_alternative}{\code{experimental_design$add_alternative()}}
\item \href{#method-set_design}{\code{experimental_design$set_design()}}
\item \href{#method-get_attributes}{\code{experimental_design$get_attributes()}}
\item \href{#method-get_design}{\code{experimental_design$get_design()}}
\item \href{#method-clone}{\code{experimental_design$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{experimental_design$new(
  alternatives = list(NULL),
  design = "random",
  n = NULL,
  identical = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_alternative"></a>}}
\if{latex}{\out{\hypertarget{method-add_alternative}{}}}
\subsection{Method \code{add_alternative()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{experimental_design$add_alternative(alternative, alternative_name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_design"></a>}}
\if{latex}{\out{\hypertarget{method-set_design}{}}}
\subsection{Method \code{set_design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{experimental_design$set_design(design = "random")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_attributes"></a>}}
\if{latex}{\out{\hypertarget{method-get_attributes}{}}}
\subsection{Method \code{get_attributes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{experimental_design$get_attributes(index = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_design"></a>}}
\if{latex}{\out{\hypertarget{method-get_design}{}}}
\subsection{Method \code{get_design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{experimental_design$get_design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{experimental_design$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
