% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial design.R
\docType{class}
\name{Factorial Design-class}
\alias{Factorial Design-class}
\alias{FactorialDesign}
\title{FactorialDesign}
\arguments{
\item{N}{The number of decision makers}

\item{p}{The number of attributes of each decision makers}

\item{J}{The number of alternatives}

\item{q}{The number of attributes of each alternatives}
}
\description{
The reference class whose methods generate a Factorial Design
}

\examples{
DM_att_names <- list("A", "B", "C"); AT_att_names <- list("D", "E", "F")
          AT_names <- list("good1", "good2"); N <- 10
          FD <- FactorialDesign(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,N=N)

}
