% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit}
\alias{fit}
\alias{loss}
\title{fit}
\usage{
loss(experimental_design, choice_set_size)
}
\arguments{
\item{experimental_design}{The experimental design tableau (should be of format long)}

\item{choice_set_size}{The size of each choice set}
}
\value{
The value of the loss function at its optimal parameter as well as the value of this optimal parameter
}
\description{
This function gives the best estimate of the preference coefficients
}
