% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{fractionalfactorialdesign}
\alias{fractionalfactorialdesign}
\title{fractionalfactorialdesign}
\usage{
fractionalfactorialdesign(A, B, U, choice_set_size, J, no_choice, nb_questions)
}
\arguments{
\item{A}{matrix of decision makers data}

\item{B}{matrix of alternatives data}

\item{U}{Utility matrix}

\item{choice_set_size}{the number of alternative per choice set}

\item{J}{number of alternatives (no_choice included)}

\item{no_choice}{TRUE FALSE indicator}

\item{nb_questions}{Number of questions to ask to each decision maker}
}
\value{
Data Frame of a random fractional factorial design
}
\description{
The method of the Experiment class which generate a full factorial design
}
