% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_beta}
\alias{gen_beta}
\title{gen_beta}
\usage{
gen_beta(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{shape1}{The first shape parameter}

\item{shape2}{The second shape parameter}

\item{ncp}{The non-centrality parameter}
}
\description{
A function generating a random matrix with beta distribution
}
\examples{
gen_beta(10, 2, param=list(shape1=2, shape2=3, ncp=0.5))

}
