% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_chi2}
\alias{gen_chi2}
\title{gen_chi2}
\usage{
gen_chi2(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{df}{Degree of freedom}

\item{ncp}{The non-centrality parameter}
}
\description{
A function generating a random matrix with chi-2 distribution
}
\examples{
gen_chi2(10, 5, param=list(df=3, ncp=1))

}
