% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_continuous_uniform}
\alias{gen_continuous_uniform}
\title{gen_continuous_uniform}
\usage{
gen_continuous_uniform(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{a}{Lower bound of the distribution's support}

\item{b}{Upper bound of the distribution's support}
}
\description{
A function generating a random matrix with continuous uniform distribution
}
\examples{
gen_continuous_uniform(10, 2, param=list(a=2, b=3))

}
