% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_discrete_uniform}
\alias{gen_discrete_uniform}
\title{gen_discrete_uniform}
\usage{
gen_discrete_uniform(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{a}{Lower bound of the distribution's support}

\item{b}{Upper bound of the distribution's support}
}
\description{
A function generating a random matrix with discrete uniform distribution
}
\examples{
gen_discrete_uniform(10, 2, param=list(a=2, b=3))

}
