% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{gen_DM_attributes}
\alias{gen_DM_attributes}
\title{gen_DM_attributes}
\arguments{
\item{law}{The name of the distribution generating the attributes}

\item{which}{The number of the columns which should be generated}

\item{nb_levels}{The vector of number of levels of the attributes designated in argument which}

\item{group}{The group to generate values. By default, every groups are concerned by the generation.}

\item{observation}{The observation function. Takes a formula as argument.}
}
\description{
The method of the Experiment class which generates some decision makers' attributes
}
\examples{
DM_att_names <- list("X1", "X2", "X3")
AT_att_names <- list("Z1", "Z2", "Z3")
AT_names <- list("good1", "good2")
groups <- c(10, 20)
FD <- Experiment(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,groups=groups)
FD$gen_DM_attributes("discrete_uniform", a=0, b=1, which="X1")
FD$gen_DM_attributes("normal", which=c("X2", "X3"), group=1)
FD$gen_DM_attributes("normal", mu=1, sd=2, which=c("X2","X3"), group=2)
FD$X

}
