% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_empirical}
\alias{gen_empirical}
\title{gen_empirical}
\usage{
gen_empirical(n, m = NULL, param)
}
\arguments{
\item{n}{The number rows}

\item{data}{Data set from which to extract the empirical distribution}
}
\description{
A function generating a random matrix from an empirical distribution
}
\examples{
gen_empirical(10, param=list(data=data.frame(S1=c(0.5, 0, 12, 6, 7.3))))

}
