% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_exp}
\alias{gen_exp}
\title{gen_exp}
\usage{
gen_exp(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{lambda}{Rate}
}
\description{
A function generating a random matrix with exponential distribution
}
\examples{
gen_exp(10, 5, param=list(lambda=3))

}
