% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_geo}
\alias{gen_geo}
\title{gen_geo}
\usage{
gen_geo(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{prob}{Probability of success}
}
\description{
A function generating a random matrix with geometric distribution
}
\examples{
gen_geo(10, 5, param=list(prob=0.5))

}
