% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_gumbel}
\alias{gen_gumbel}
\title{gen_gumbel}
\usage{
gen_gumbel(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{location}{The optional location parameter}

\item{scale}{The optional scale, a positive number}
}
\description{
A function generating a random matrix with gumbel distribution
}
\examples{
gen_gumbel(10, 5, param=list(location=3))
gen_gumbel(10, 5, param=list(location=3, scale=4))

}
