% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_normal}
\alias{gen_normal}
\title{gen_normal}
\usage{
gen_normal(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{mu}{The mean vector}

\item{sd}{The optional optional covariance matrix}
}
\description{
A function generating a random matrix with multivariate normal distribution
}
\examples{
gen_normal(10, 2, param=list(mu=rep(-10,2)))

}
