% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_poisson}
\alias{gen_poisson}
\title{gen_poisson}
\usage{
gen_poisson(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{lambda}{rate}
}
\description{
A function generating a random matrix with Poisson distribution
}
\examples{
gen_poisson(10, 5, param=list(lambda=3))

}
