% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{gen_preference_coefficients}
\alias{gen_preference_coefficients}
\title{gen_preference_coefficients}
\arguments{
\item{which}{The number of the columns which should be generated}
}
\description{
The method of the Experiment class which generates some preference coefficients' attributes
}
\examples{
DM_att_names <- list("S1", "S2", "S3")
AT_att_names <- list("X1", "X2", "X3")
AT_names <- list("good1", "good2")
groups <- c(10, 20)
FD <- Experiment(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,groups=groups)
FD$gen_preference_coefficients("student", heterogeneity=TRUE, location=-100,  scale=1, df=4, which=c(1:4), group=1)
FD$gen_preference_coefficients("student", heterogeneity=TRUE, location=100,  scale=1, df=4, which=c(1:4), group=2)
FD$gen_preference_coefficients("normal", heterogeneity=TRUE, mu=0, sd=2, which=5)
FD$gen_preference_coefficients("discrete_uniform", heterogeneity=TRUE, a=1, b=5, which=6)
FD$beta

}
