% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gen_student}
\alias{gen_student}
\title{gen_student}
\usage{
gen_student(n, m, param)
}
\arguments{
\item{n}{The number rows}

\item{m}{The number of columns}

\item{location}{The optional location parameter}

\item{scale}{The optional scale matrix}
}
\description{
A function generating a random matrix with multivariate student distribution
}
\examples{
gen_student(10, 2, param=list(scale=diag(3,2), df=5))

}
