% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\docType{class}
\name{individual}
\alias{individual}
\title{Individual class}
\description{
The `individual` R6 class serves to define individuals' profiles to be included into  population.
The object stores information about the individuals' characteristics, as well as the decision rules.
The values may be specified as random declaration using statistical functions of `r` type.
The object is initialized without attributes and should be later populated with them.

Append a list of `call` attributes definitions to the attributes.
The attributes should have a name and generation procedure defined (ex: `Age = rnorm(mean = 40, sd = 10)`)

Create or replace an existing decision rule

Get a vector of available characteristics' names

Get a list of alternatives' generation rules

Extract `decision_rule` object
}
\section{Modifying methods}{
NA
}

\section{Querrying methods}{
NA
}

\examples{
ind <- individual$new()
ind$add_characteristics(Age = rnorm(mean = 40, sd = 10))
ind$add_decision_rule(drule <- decision_rule$new())
ind$get_char()
ind$get_laws()
ind$get_rule()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{characteristics}}{A list of characteristics' definitions}

\item{\code{decision_rule}}{A decision rule object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_characteristics}{\code{individual$add_characteristics()}}
\item \href{#method-add_decision_rule}{\code{individual$add_decision_rule()}}
\item \href{#method-get_chars}{\code{individual$get_chars()}}
\item \href{#method-get_laws}{\code{individual$get_laws()}}
\item \href{#method-get_rule}{\code{individual$get_rule()}}
\item \href{#method-clone}{\code{individual$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-add_characteristics}{}}}
\subsection{Method \code{add_characteristics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{individual$add_characteristics(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_decision_rule"></a>}}
\if{latex}{\out{\hypertarget{method-add_decision_rule}{}}}
\subsection{Method \code{add_decision_rule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{individual$add_decision_rule(decision_rule)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_chars"></a>}}
\if{latex}{\out{\hypertarget{method-get_chars}{}}}
\subsection{Method \code{get_chars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{individual$get_chars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_laws"></a>}}
\if{latex}{\out{\hypertarget{method-get_laws}{}}}
\subsection{Method \code{get_laws()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{individual$get_laws()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rule"></a>}}
\if{latex}{\out{\hypertarget{method-get_rule}{}}}
\subsection{Method \code{get_rule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{individual$get_rule()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{individual$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
