% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{infoDesign}
\alias{infoDesign}
\title{infoDesign}
\usage{
infoDesign(
  name,
  experimental_design,
  AT_names,
  choice_set_size,
  J,
  no_choice,
  DM_att_names,
  AT_att_names
)
}
\arguments{
\item{name}{The name of the chosen design}

\item{experimental_design}{A generated experimental design}

\item{choice_set_size}{the number of alternative per choice set}

\item{J}{number of alternatives (no_choice included)}

\item{no_choice}{TRUE FALSE indicator}

\item{DM_att_names}{The names of the decision makers' attributes}

\item{AT_att_names}{The names of the alternatives' attributes}
}
\description{
A function which provides a summary about an experimental design
}
