% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{InfoMatrix}
\alias{InfoMatrix}
\title{InfoMatrix}
\usage{
InfoMatrix(experimental_design)
}
\arguments{
\item{experimental_design}{The table representation of an experimental design}
}
\description{
The function which computes the information matrix.
}
