% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative.R
\name{is.alternative}
\alias{is.alternative}
\title{Alternative class testing}
\usage{
is.alternative(alternative)
}
\arguments{
\item{alternative}{Input object to test}
}
\value{
Logic
}
\description{
Test if the given object has an `alternative` class.
}
\examples{
alt <- alternative$new()
is.alternative(alt)
}
