% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_rule.R
\name{is.decision_rule}
\alias{is.decision_rule}
\title{Decision rule class testing}
\usage{
is.decision_rule(decision_rule)
}
\arguments{
\item{decision_rule}{Input object to test}
}
\value{
Logic
}
\description{
Test if the given object has an `decision_rule` class.
}
\examples{
drule <- decision_rule$new()
is.decision_rule(drule)
}
