% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_design.R
\name{is.experimental_design}
\alias{is.experimental_design}
\title{Experimental design class testing}
\usage{
is.experimental_design(experimental_design)
}
\arguments{
\item{experimental_design}{Input object to test}
}
\value{
Logic
}
\description{
Test if the given object has an `experimental_design` class.
}
\examples{
edesign <- experimental_design$new()
is.population(edesign)

}
