% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{is.individual}
\alias{is.individual}
\title{Individual class testing}
\usage{
is.individual(individual)
}
\arguments{
\item{individual}{Input object to test}
}
\value{
Logic
}
\description{
Test if the given object has an `individual` class.
}
\examples{
ind <- individual$new()
is.individual(ind)
}
