% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{is.population}
\alias{is.population}
\title{Population class testing}
\usage{
is.population(population)
}
\arguments{
\item{population}{Input object to test}
}
\value{
Logic
}
\description{
Test if the given object has an `population` class.
}
\examples{
pop <- population$new()
is.population(pop)
}
