% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{long_format}
\alias{long_format}
\title{long_format}
\usage{
long_format(
  design_expe,
  characterisnics_names,
  attributes_names,
  choice_set_size
)
}
\arguments{
\item{design_expe}{An experimental design of format wide}
}
\value{
The long format of a wide format experimental design
}
\description{
This function transforms a wide format experimental design into a long format design
}
