% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{mixed_design}
\alias{mixed_design}
\title{Mixed experimental design}
\usage{
mixed_design(experimental_design, n)
}
\arguments{
\item{experimental_design}{Input experimnetal design object}

\item{n}{The n is not used in current version}
}
\value{
data.frame A dataset of choice sets (Z) respecting FF Design.
}
\description{
Create mixed experimental design.
The design configuration assumes that alternatives' attributes are mixed:
vectors declared using `c()` assignement and
`declared using `r` class simulation functions.
}
\examples{
# Create alternatives
alt1 <- alternative$new()
alt1$add_attributes(Quality = c(0:1), Price = rnorm(mean = 2, sd = 1))
alt2 <- alternative$new()
alt2$add_attributes(Size = c(1:3), Price = runif(min = 1, max = 3))

# Regroup alternatives into design
edesign <- experimental_design$new(alternatives = list(alt1, alt2), design = "mixed")
Z <- random_design(edesign)
}
