% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\docType{class}
\name{noise-class}
\alias{noise-class}
\alias{noise}
\title{ob_decision_makers_att}
\arguments{
\item{p}{The number of attributes of each decision makers}

\item{N}{The number of decision makers}

\item{J}{The number of attributes of alternatives}
}
\description{
The class of observed decision makers' attributes

The class noise
}

\examples{
N <- 10; p <- 3; a <- ob_decision_makers_att$new(N=N, p=p);
          a$gen("student", param=list(location=rep(100, 3), df=3)); a$gen(); a$gen("help")

N <- 10; J <- 5; a <- noise$new(N=N, J=J);
          a$gen()
          a$gen("normal", param=list(mu=rep(-1000, J)))
          a$gen("student", param=list(scale=diag(2, J),df=10))
          a$gen("discrete_uniform")

}
