% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalization}
\alias{normalization}
\title{normalization}
\usage{
normalization(V, U)
}
\arguments{
\item{V}{The matrix of representative utilities}

\item{U}{The matrix of the utilities}
}
\description{
The function which normalizes the representative utility
}
\examples{
V <- matrix(rnorm(30), 10, 3)
U <- matrix(rnorm(30), 10, 3) + matrix(rnorm(30), 10, 3)
normalization(V, U)

}
