% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population_gen}
\alias{population_gen}
\title{Generate population}
\usage{
population_gen(population, seed = NULL, class = TRUE)
}
\arguments{
\item{population}{Input population configuration}

\item{seed}{The seed to be set before attempting to generate population data.
Defaults to NULL.}

\item{class}{Logic.
Indicates whether or not to include individual class information to the resulting data.frame.
Defaults to NULL.}
}
\value{
data.frame A data.frame (X) with simulated population (one row per individual).
}
\description{
Generate population data (X in standart notation) from a population object.
}
\examples{
# Create individuals
ind1 <- individual$new()
ind1$add_characteristics(Age = rnorm(mean = 40, sd = 10))
ind1$add_decision_rule(drule <- decision_rule$new())
ind2 <- individual$new()
ind2$add_characteristics(Age = rnorm(mean = 30, sd = 5))
ind2$add_decision_rule(drule <- decision_rule$new())

# Regroup individuals into population
pop <- population$new(profiles = list(ind1, ind2), n = list(10, 15))
X <- population_gen(pop)
}
