% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\docType{class}
\name{Preference coefficients-class}
\alias{Preference coefficients-class}
\alias{preference_coef}
\title{preference_coef}
\arguments{
\item{N}{The number of decision makers}

\item{p}{The number of attributes of each decision makers}

\item{q}{The number of attributes of each attributes}
}
\description{
Class of preference coefficients
}

\examples{
S <- matrix(rnorm(100), nrow = 20, byrow = TRUE)
X <- matrix(rnorm(20), nrow = 4, byrow = TRUE)
choice <- c(rep(1,5), rep(2,5), rep(3,5), rep(4,5))
estimator <- preference_coef(N=20, p=5, q=5)

a <- preference_coef(N=20, p=5, q=5); a$gen(law="student"); a$gen(heterogeneity=TRUE)

}
