% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{random_design}
\alias{random_design}
\title{Randomized experimental design}
\usage{
random_design(experimental_design, n)
}
\arguments{
\item{experimental_design}{Input experimnetal design object}

\item{n}{The number of choice sets to be generated.}
}
\value{
data.frame A dataset of choice sets (Z) respecting randomized Design.
}
\description{
Create Fully Randomized experimental design.
The design configuration assumes that all alternatives' attributes are defined by random generators:
`declared using `r` class simulation functions.
Other attribute types will be ignored.
}
\examples{
# Create alternatives
alt1 <- alternative$new()
alt1$add_attributes(Quality = purrr::rbernoulli(p = 0.6), Price = rnorm(mean = 5))
alt2 <- alternative$new()
alt2$add_attributes(Size = runif(min = 0, max = 1), Price = rnorm(mean = 6))

# Regroup alternatives into design
edesign <- experimental_design$new(alternatives = list(alt1, alt2))
Z <- random_design(edesign, n = 1)
}
