% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial design.R
\name{representative_utility}
\alias{representative_utility}
\title{representative_utility}
\arguments{
\item{funct}{By default, the representative utility function is the sum of the alternatives of
a decision maker and the alternatives of an alternative weighted by the preference coefficients.
However, the user can choose whatever function with whatever coefficients}
}
\description{
The method of the Factorial design class which computes the representative utility
}
\examples{
N <- 10; AT_names <- list("good1", "good2")
DM_att_names <- list("A", "B", "C")
AT_att_names <- list("D", "E", "F")
FD <- FactorialDesign(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,N=N)
FD$gen_DM_attributes()
FD$gen_AT_attributes()
FD$gen_preference_coefficients()
FD$representative_utility()
FD$V

}
