% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample.R
\name{rsample}
\alias{rsample}
\title{Random sampling}
\usage{
rsample(x, n, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{A vector of one or more elements from which to choose.}

\item{n}{A non-negative integer giving the number of items to choose.}

\item{replace}{Logical, whether the sampling should be with replacement.}

\item{prob}{A vector of probability weights for obtaining the elements of the vector being sampled.}
}
\value{
vector of length `n` containing elements from `x`.
}
\description{
This function is a wrapper around the `base::sample()`.
For more information refer to `?base::sample()`.
}
\examples{
x <- 1:10; n <- 5
rsample(x, n, replace = FALSE)
}
