% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial design.R
\name{total_utility}
\alias{total_utility}
\title{total_utility}
\description{
The method of the Factorial design class which generates the utility of the agents for each good
}
\examples{
N <- 10
AT_names <- list("good1", "good2")
DM_att_names <- list("A", "B", "C")
AT_att_names <- list("D", "E", "F")
FD <- FactorialDesign(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,N=N)
FD$gen_AT_attributes()
FD$gen_DM_attributes()
FD$gen_preference_coefficients()
FD$representative_utility()
FD$total_utility()
FD$Epsilon; FD$U; FD$choice_order
FD$total_utility("student", mu=3, df=4) # Some examples
FD$total_utility("discrete_uniform") # Some examples
FD$Epsilon
FD$U
FD$choice_order
FD$Epsilon
FD$U
FD$choice_order

}
