% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_helpers.R
\name{check_attributes}
\alias{check_attributes}
\title{Attributes check}
\usage{
check_attributes(Z, attributes)
}
\arguments{
\item{Z}{Input data.frame Z}

\item{attributes}{Desired vector of attributes to be present in Z.}
}
\value{
Z With missing attributes filled with NA's.
}
\description{
Verification that all desired characteristics are present in $Z$.
This is an internal function which serves to simplify concatenation of multiple $Z$ when alternatives have different attributes.
}
\examples{
# Create alternatives
alt1 <- alternative$new()
alt1$add_attributes(Quality = runif(min = 0, max = 1), Price = rnorm(mean = 5))
alt2 <- alternative$new()
alt2$add_attributes(Size = runif(min = 0, max = 1), Price = rnorm(mean = 6))

# Regroup alternatives into design
edesign <- experimental_design$new(alternatives = list(alt1, alt2))
Z <- random_design(edesign, n = 1)

# Querry object
attr <- edesign$get_attributes()
Z <- check_attributes(Z, attr)
}
