% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_helpers.R
\name{class_laws}
\alias{class_laws}
\title{Classify generating laws}
\usage{
class_laws(laws)
}
\arguments{
\item{laws}{A list of generation rules (rules) as provided by single alternative.}
}
\value{
character A vector of classes in character format.
}
\description{
This is an internal function which classifies the attribute / characteristic generating functions.
The function serves to distinguish between the different input formats.
This ensures flexibility of the attributes' specifications.
}
\examples{
# Create alternative
alt1 <- alternative$new()
alt1$add_attributes(Quality = runif(min = 0, max = 1), Price = rnorm(mean = 5))

# Class laws
claws <- class_laws(alt1$get_laws())

}
