% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_random}
\alias{compose_random}
\title{Compose random choice sets}
\usage{
compose_random(experimental_design, n = NULL, size)
}
\arguments{
\item{experimental_design}{An `experimental_design` object.}

\item{n}{Number of different choice situations per individual.
Is NULL by default, ensuring that the function will retreive the n value from the `experimental_design`.}

\item{size}{Population size for Z to be adjusted to.}
}
\value{
data.frame Z dataset
}
\description{
This is an internal function which creates distinct choice sets configurations for give experimental design.
The function simplifies the size adjustment of Z depending on the population size of X.
Each individuals faces a personalised Z version.
}
\examples{
# Create alternatives
alt1 <- alternative$new()
alt1$add_attributes(Quality = runif(min = 0, max = 1), Price = rnorm(mean = 5))
alt2 <- alternative$new()
alt2$add_attributes(Size = runif(min = 0, max = 1), Price = rnorm(mean = 6))

# Regroup alternatives into design
edesign <- experimental_design$new(alternatives = list(alt1, alt2), n = 4)

# Compose
Z <- compose_random(edesign, size = 2)
}
