% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Dcriterion}
\alias{Dcriterion}
\title{Dcriterion}
\usage{
Dcriterion(experimental_design)
}
\arguments{
\item{experimental_design}{The design matrix}
}
\description{
The function which computes the determinant of the information matrix.
}
