% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{design}
\alias{design}
\title{design}
\arguments{
\item{name}{The name of the experimental design to implement.}

\item{choice_set_size}{The number of alternative per choice set.}

\item{clustered}{Determines the way the data is represented in the experimental design table.
0 if the matrices of decision makers' characteristics and alternatives' attributes are row.
1 if the matrices of decision makers' characteristics and alternatives' attributes are clustered.
2 if the matrices of decision makers' characteristics and alternatives' attributes are categorized.}

\item{format}{If "long" (default) the design is expressed in long format and wide format otherwise.}
}
\description{
The method of the Experiment class which generate a full factorial design
}
\examples{
DM_att_names <- list("S1", "S2", "S3")
AT_att_names <- list("X1", "X2", "X3")
AT_names <- list("good1", "good2", "good3", "good4")
groups <- c(10, 20)
FD <- Experiment(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names, groups=groups)
FD$gen_AT_attributes()
FD$gen_DM_attributes()
FD$gen_preference_coefficients()
FD$utility()
FD$design(choice_set_size=2, clustered=0)
FD$design(choice_set_size=2, clustered=1, nb_levels_DM=c(3,3,4), nb_levels_AT=c(3, 2, 2))
FD$design(choice_set_size=2, clustered=2, nb_levels_DM=c(3,3,4), nb_levels_AT=c(3, 2, 2))
}
