% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_compose.R
\name{experiment_compose}
\alias{experiment_compose}
\title{Compose experiment}
\usage{
experiment_compose(population, experimental_design, seed = NULL)
}
\arguments{
\item{population}{A `population` object.}

\item{experimental_design}{An `experimental_desing` object.}

\item{seed}{The desired seed to be set before data generation.
No seed is set by default (`seed = NULL`).}
}
\value{
data.frame A complete experimental dataset XZ
}
\description{
This function simulates the data preparation step for an experiment.
The function performs coupling between X (population) and Z (experimental_design) datasets.
}
\examples{
# Generate individual profile and population
ind3 <- individual$new()
ind3$add_characteristics(Age = rnorm(mean = 50, sd = 4), Salary = runif(min = 1, max = 5))
ind3$add_decision_rule(drule <- decision_rule$new())
pop <- population$new(profiles = list(ind3), n = list(5))

# Create alternatives and regroup them into design
alt1 <- alternative$new()
alt1$add_attributes(Quality = runif(min = 0, max = 1), Price = rnorm(mean = 5))
alt2 <- alternative$new()
alt2$add_attributes(Size = runif(min = 0, max = 1), Price = rnorm(mean = 6))
edesign <- experimental_design$new(alternatives = list(alt1, alt2), n = 4)

# Compose dataset
XZ <- experiment_run(pop, edesign, seed = 10)
}
