% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_run.R
\name{experiment_run}
\alias{experiment_run}
\title{Run experiment}
\usage{
experiment_run(population, experimental_design, seed = NULL)
}
\arguments{
\item{population}{A `population` object.}

\item{experimental_design}{1n `experimental_desing` object.}
}
\value{
data.frame A complete experimental dataset with associated utilities.
}
\description{
This function simulates running of an experiment.
First, the `experiment_compose` is executed to obtain a coupling between X and Z matrices.
Then Deterministic Utilities (DU)  and Total Utilities (TU) are calculated using the specified formulas.
Note, that even the TU abbreviation is used this still makes possible to recalibrate the generation functions, tranforms and decision tules to create a Regret Minimisation (RM) framework i.e.
}
\examples{
res <- experiment_run(population, experimental_design)
}
