% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{fullfactorialdesign}
\alias{fullfactorialdesign}
\title{fullfactorialdesign}
\usage{
fullfactorialdesign(X, Z, U, choice_set_size, J, no_choice, format = "long")
}
\arguments{
\item{X}{matrix of decision makers data}

\item{Z}{matrix of alternatives data}

\item{U}{Utility matrix}

\item{choice_set_size}{the number of alternative per choice set}

\item{J}{number of alternatives (no_choice included)}

\item{no_choice}{TRUE FALSE indicator}
}
\description{
The method of the Experiment class which generate a full factorial design
}
