% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{gen_AT_attributes}
\alias{gen_AT_attributes}
\title{gen_AT_attributes}
\arguments{
\item{law}{The name of the distribution generating the attributes}

\item{which}{The number of the columns which should be generated}

\item{nb_levels}{The vector of number of levels of the attributes designated in argument which}
}
\description{
The method of the Experiment class which generates some alternatives' attributes
}
\examples{
DM_att_names <- list("S1", "S2", "S3")
AT_att_names <- list("X1", "X2", "X3")
AT_names <- list("good1", "good2", "good3", "good4")
groups <- c(10, 20)
FD <- Experiment(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names, groups=groups)
FD$gen_AT_attributes("discrete_uniform", b=2, which=c("X1","X2"))
FD$gen_AT_attributes("normal", which=c("X3"), nb_levels=3)
FD$X; FD$X_clustered; FD$X_category

}
