% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial design.R
\name{gen_DM_attributes}
\alias{gen_DM_attributes}
\title{gen_DM_attributes}
\arguments{
\item{law}{The name of the distribution generating the attributes}

\item{which}{The number of the columns which should be generated}
}
\description{
The method of the Factorial design class which generates some decision makers' attributes
}
\examples{
N <- 10; AT_names <- list("good1", "good2")
DM_att_names <- list("A", "B", "C")
AT_att_names <- list("D", "E", "F")
FD <- FactorialDesign(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,N=N)
FD$gen_DM_attributes("discrete_uniform", a=0, b=1, which="A")
FD$gen_DM_attributes("normal", which=c("B","C"))
FD$gen_DM_attributes("normal", mu=1, sd=2, which=c("B","C"))
FD$S

}
