% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial design.R
\name{gen_preference_coefficients}
\alias{gen_preference_coefficients}
\title{gen_preference_coefficients}
\arguments{
\item{which}{The number of the columns which should be generated}
}
\description{
The method of the Factorial design class which generates some preference coefficients' attributes
}
\examples{
N <- 10
AT_names <- list("good1", "good2")
DM_att_names <- list("A", "B", "C")
AT_att_names <- list("D", "E", "F")
FD <- FactorialDesign(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,N=N)
FD$gen_preference_coefficients("student", heterogeneity=TRUE, location=-100,  scale=1, df=4, which=c(1:4))
FD$gen_preference_coefficients("normal", heterogeneity=TRUE, mu=0, sd=2, which=c(5))
FD$gen_preference_coefficients("discrete_uniform", heterogeneity=TRUE, a=1, b=5, which=6)
FD$beta

}
