% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{generation}
\alias{generation}
\title{generation}
\usage{
generation(law, n, m, param)
}
\arguments{
\item{law}{The distribution generating the matrix}

\item{n}{The number rows}

\item{m}{The number of columns}
}
\description{
3 distributions are currently available: normal, student,
discrete, gumbel. In addition, gen("hepl") provides information about how to call
the optional parameters
}
\examples{
generation("student", 10, 5, param=list(location=rep(100, 5)))
generation("help")

}
