% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial design.R
\name{map}
\alias{map}
\title{map}
\description{
The method of the Factorial design class which draws a preferenc mapping
}
\examples{
N <- 10
AT_names <- list("good1", "good2")
DM_att_names <- list("A", "B", "C")
AT_att_names <- list("D", "E", "F")
FD <- FactorialDesign(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,N=N)
FD$factorial_design()
FD$map("A", "D")

}
