% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{representative_utility}
\alias{representative_utility}
\title{representative_utility}
\description{
The method of the Experiment class which computes the representative utility
}
\examples{
DM_att_names <- list("S1", "S2", "S3")
AT_att_names <- list("X1", "X2", "X3")
AT_names <- list("good1", "good2")
groups <- c(10, 20)
FD <- Experiment(DM_att_names=DM_att_names, AT_att_names=AT_att_names, AT_names=AT_names,groups=groups)
FD$gen_DM_attributes()
FD$gen_AT_attributes()
FD$gen_preference_coefficients()
FD$representative_utility()
FD$V

}
