% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary}
\alias{summary}
\alias{summary.Exepriment}
\title{summary}
\usage{
\method{summary}{Exepriment}(FD)
}
\arguments{
\item{FD}{An instance of the class Experiment}
}
\value{
Some information about an instance of the class Experiment
}
\description{
This function returns a summary of the class Experiment
}
